Convert AccordionItemTitle.wrapper to tsx
